requirements.txt
pandas
matplotlib
numpy
scikit-learn
seaborn
requests
flask
pymongo
sqlalchemy
pytest
imblearn
xgboost
catboost
from_root
fastapi
uvicorn
jinja2
pydantic
dill
boto3
python-dotenv
neuro_mf
botocore
python-multipart
evidently
mypy-boto3-s3

-e .